requirements.txt
alembic==1.8.1
autopep8==2.0.0
Bootstrap-Flask==2.1.0
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
psycopg2==2.9.5
pycodestyle==2.9.1
python-dotenv==0.21.0
SQLAlchemy==1.4.44
tomli==2.0.1
Werkzeug==2.2.2
WTForms==3.0.1
